+++ src/services/branch-manager.js
@@ -59,16 +59,6 @@ function pushBranch(destinationEnvironmentName, environmentSecret) {
     .send({ destinationEnvironmentName: encodeURIComponent(destinationEnvironmentName) });
 }
 
-function deploy({ id }, environmentSecret) {
-  const authToken = authenticator.getAuthToken();
-
-  return agent
-    .post(`${serverHost()}/api/layout/deploy`)
-    .set('Authorization', `Bearer ${authToken}`)
-    .set('forest-secret-key', `${environmentSecret}`)
-    .send({ environmentId: id });
-}
-
 function switchBranch({ id }, environmentSecret) {
   const authToken = authenticator.getAuthToken();
 
@@ -120,6 +110,5 @@ module.exports = {
   createBranch,
   pushBranch,
   switchBranch,
-  deploy,
   handleBranchError,
 };
